Fixed Uped the Porla models

+++ src/NzbDrone.Core/Download/Clients/Porla/Models/PorlaSessionList.cs
@@ -3,12 +3,14 @@ using Newtonsoft.Json;
 
 namespace NzbDrone.Core.Download.Clients.Porla.Models
 {
+    /// <summary> Implementation of the <em>session</em> field in the response from <a href="https://github.com/porla/porla/blob/v0.37.0/src/methods/sessions/sessionslist.cpp">sessionslist.cpp</a> data type </summary>
     public sealed class ResponsePorlaSessionList
     {
         [JsonProperty("sessions", NullValueHandling = NullValueHandling.Ignore)]
         public ReadOnlyCollection<PorlaSession> Sessions { get; set; }
     }
 
+    /// <summary> Implementation of the session data type from the response <a href="https://github.com/porla/porla/blob/v0.37.0/src/methods/sessions/sessionslist.cpp">sessionslist.cpp</a></summary>
     public class PorlaSession
     {
         [JsonProperty("is_dht_running", NullValueHandling = NullValueHandling.Ignore)]
@@ -24,6 +26,6 @@ namespace NzbDrone.Core.Download.Clients.Porla.Models
         public string Name { get; set; }
 
         [JsonProperty("torrents_total", NullValueHandling = NullValueHandling.Ignore)]
-        public long TorrentsTotal { get; set; }
+        public int TorrentsTotal { get; set; }
     }
 }
